minot

+++ examples/improved_wgan_AC.py
@@ -281,8 +281,8 @@ generator_layers = generator(generator_input)
 discriminator_layers_for_generator = discriminator(generator_layers)
 
 generator_model = Model(inputs=[generator_input],
-                        outputs=[K.cast(discriminator_layers_for_generator[0]),
-                                 K.cast(discriminator_layers_for_generator[1:])])
+                        outputs=[K.cast(discriminator_layers_for_generator[0],  dtype = 'float32'),
+                                 K.cast(discriminator_layers_for_generator[1:], dtype = 'float32')])
 # We use the Adam paramaters from Gulrajani et al.
 generator_model.compile(optimizer=Adam(0.0001, beta_1=0.5, beta_2=0.9), loss=[wasserstein_loss,
                                                                               label_loss])
